Ensure that wine products aren't automatically publishsed
