+++ src/Tasks/Services/StockImport.php
@@ -3,16 +3,13 @@
 
 namespace Isobar\Flow\Tasks\Services;
 
-use BadMethodCallException;
 use Exception;
 use Isobar\Flow\Config\FlowConfig;
 use Isobar\Flow\Exception\FlowException;
 use Isobar\Flow\Services\FlowAPIConnector;
 use Isobar\Flow\Services\Product\StockAPIService;
 use SilverStripe\Control\Director;
-use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\ValidationException;
-use SilverStripe\Versioned\ChangeSet;
 use SwipeStripe\Common\Product\ComplexProduct\ComplexProductVariation;
 
 /**
@@ -21,33 +18,9 @@ use SwipeStripe\Common\Product\ComplexProduct\ComplexProductVariation;
  */
 class StockImport
 {
+    use ChangeSetPublishTrait;
+
     protected $threshold;
-
-    /**
-     * @var ChangeSet
-     */
-    protected $changeSet = null;
-
-    /**
-     * Publish / update a record
-     *
-     * @param DataObject $record
-     * @throws ValidationException
-     */
-    protected function publish(DataObject $record)
-    {
-        if (empty($this->changeSet)) {
-            throw new BadMethodCallException("runImport not called");
-        }
-        if ($record->isChanged(null, DataObject::CHANGE_VALUE)) {
-            $record->write();
-        }
-        // Ensure changeset is saved when adding the first item
-        if (!$this->changeSet->isInDB()) {
-            $this->changeSet->write();
-        }
-        $this->changeSet->addObject($record);
-    }
 
     /**
      * Imports data from Flow XML feed
@@ -63,16 +36,12 @@ class StockImport
 
         // import PIMS data to temp table
         try {
-            $this->changeSet = ChangeSet::create();
+            $this->initChangeSet();
             $this->importData();
         } catch (Exception $e) {
             throw new FlowException($e->getMessage(), $e->getCode());
         } finally {
-            // If we have saved at least one record, publish changeset
-            if ($this->changeSet->isInDB()) {
-                $this->changeSet->publish();
-            }
-            $this->changeSet = null;
+            $this->finishChangeSet();;
         }
 
         if (Director::is_cli()) {
